테마마켓에서 테마 내리기

+++ core/src/main/kotlin/theme/service/TimetableThemeService.kt
@@ -6,6 +6,8 @@ import com.wafflestudio.snutt.common.exception.DuplicateThemeNameException
 import com.wafflestudio.snutt.common.exception.InvalidThemeColorCountException
 import com.wafflestudio.snutt.common.exception.InvalidThemeTypeException
 import com.wafflestudio.snutt.common.exception.NotDefaultThemeErrorException
+import com.wafflestudio.snutt.common.exception.NotPublishedThemeException
+import com.wafflestudio.snutt.common.exception.PublishedThemeDeleteErrorException
 import com.wafflestudio.snutt.common.exception.ThemeNotFoundException
 import com.wafflestudio.snutt.friend.dto.FriendState
 import com.wafflestudio.snutt.friend.service.FriendService
@@ -57,6 +59,11 @@ interface TimetableThemeService {
     ): TimetableTheme
 
     suspend fun deleteTheme(
+        userId: String,
+        themeId: String,
+    )
+
+    suspend fun deletePublishedTheme(
         userId: String,
         themeId: String,
     )
@@ -244,7 +251,24 @@ class TimetableThemeServiceImpl(
         }
         timetableRepository.saveAll(timetables).collect()
 
+        if (theme.status == ThemeStatus.PUBLISHED) throw PublishedThemeDeleteErrorException
+
         timetableThemeRepository.delete(theme)
+    }
+
+    override suspend fun deletePublishedTheme(
+        userId: String,
+        themeId: String,
+    ) {
+        val theme = getCustomTheme(userId, themeId)
+
+        if (theme.status == ThemeStatus.PUBLISHED) {
+            theme.status = ThemeStatus.PRIVATE
+        } else {
+            throw NotPublishedThemeException
+        }
+
+        timetableThemeRepository.save(theme)
     }
 
     override suspend fun copyTheme(
